+++ projects/ez-vid-ang/src/lib/controls/audio-track-selector/audio-track-selector.spec.ts
@@ -1,0 +1,7 @@
+import { EvaAudioTrackSelector } from './audio-track-selector';
+
+describe('EvaAudioTrackSelector', () => {
+  it('should be defined', () => {
+    expect(EvaAudioTrackSelector).toBeDefined();
+  });
+});